Move MockSA from WFE to mocks.
